Metabolite metadata in negative polarity for GNPS mgf export

This data frame contains all the metadata required in the MS2extract package to export the library in a GNPS .mgf friendly format.

Usage

metdt_GNPS_neg

Format

metdt_GNPS_neg

A data frame containing 168 observations (metabolites) and eight columns

COMPOUND_NAME

character, compound name

INSTRUMENT

character, instrument type QTOF as qTof

COLLISIONENERGY

character, collision energy applied in MS/MS

IONSOURCE

character, ion source LC-ESI in our case

SMILES

character, smiles of the metabolite

INCHI

character, INCHIKEY

INCHIAUX

character, auxilar INCHIKEY

IONMODE

character, ionization polarity

PUBMED

character, PUBMED identifier

ACQUISITION

character, source of the compound

DATACOLLECTOR

character, data collector

INTEREST

character, interest of the compound

LIBQUALITY

integer, quality of the library (1 - Gold, 2 - Silver, 3 - Bronze)

GENUS

character, genus

SPECIES

character, species

STRAIN

character, strain

CASNUMBER

numeric, CAS number

PI

character, principal investigator

Details

Be aware this table differs from the GNPS upload template, and the final table in a .tsv format is created while using MS2extract write_mgf_gnps() function.