Metabolite library metadata in negative polarity at 40 eV second batch

This data frame contains the metabolite metadata to be included in the final library file.

Usage

metdt_pos40_6546

Format

metdt_pos40_6546

NAME

character, metabolite name

FORMULA

character, chemical formula

INCHIKEY

character, inchikey recovered form PUBCHEM

SMILES

character, canonical smiles recovered from PUBCHEM

PRECURSORTYPE

character, precursor type [M+H]+ for postive and [M-H]- for negative polarity

COLLISIONENERGY

character, collision energy used for MS/MS

INSTRUMENTTYPE

character, intrument type used for data collection

IONMODE

character, Positive or Negative, restricted characters

Details

This specific data frame contains the metabolite information that was captured in positive polarity at 40 eV (CE) in the Agilent QTOF 6546.