Metabolite information to batch read in Negative polarity at 20 eV CE

This data frame contains the basic info required to batch import multiple files to create in-house libraries by using the MS2extract package.

Usage

read_neg20_6546

Format

read_neg20_6546

A data frame containing 58 observations (metabolites) and 6 columns

Name

character, metabolite name

Formula

character, chemical formula of the metabolite

Ionizatioin_move

character, the polarity (Positive or Negative)

min_rt

integer, minimum retention time of metabolite elution (seconds)

max_rt

integer, maximum retention time of metabolite elution (seconds)

File

character, filename not including the path

Details

Metabolites included in this list were analyzed by targeted methods using an Agilent 6546 QTOF.